Moved type into Person
